pyproject.toml
[project]
name = "ai-agent-recruitment-agency"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.177.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.0.11",
    "langchain-core>=0.3.68",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.0.10",
    "langchain-groq>=0.3.6",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.2",
    "pydantic[email]>=2.11.7",
    "pymupdf>=1.26.3",
    "pypdf>=5.8.0",
    "pytz>=2025.2",
    "qwen-api>=1.4.13",
    "sentence-transformers>=5.0.0",
]
